Исправленный вариант

+++ src/main/java/courses/PP_Spring/dao/UserDaoImpl.java
@@ -3,6 +3,7 @@ package courses.PP_Spring.dao;
 
 import courses.PP_Spring.model.User;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.PersistenceContext;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,7 +16,7 @@ import java.util.List;
 public class UserDaoImpl implements UserDao {
 
     @PersistenceContext
-    EntityManager entityManager;
+    private EntityManager entityManager;
 
     @Override
     public void save(User user) {
@@ -23,24 +24,36 @@ public class UserDaoImpl implements UserDao {
     }
 
     @Override
-    @Transactional(readOnly = true)
     public List<User> allUsers() {
         return entityManager.createQuery("select u from courses.PP_Spring.model.User u", User.class).getResultList();
     }
 
     @Override
     public User update(User user) {
-        return entityManager.merge(user);
+        if (getById(user.getId()) != null) {
+            return entityManager.merge(user);
+        } else {
+            throw new EntityNotFoundException("Пользователь с таким ID " + user.getId() + " не найден");
+        }
+
     }
 
     @Override
     public void delete(User user) {
-        entityManager.remove(entityManager.contains(user) ? user : entityManager.merge(user));
+        if (getById(user.getId()) != null) {
+            entityManager.remove(entityManager.contains(user) ? user : entityManager.merge(user));
+        } else {
+            throw new EntityNotFoundException("Пользователь с таким ID " + user.getId() + " не найден");
+        }
+
     }
 
     @Override
-    @Transactional(readOnly = true)
     public User getById(Integer id) {
-        return entityManager.find(User.class, id);
+        User user = entityManager.find(User.class, id);
+        if (user == null) {
+            throw new EntityNotFoundException("Пользователь с таким ID " + id + " не найден");
+        }
+        return user;
     }
 }
